remove spack and spack manager classes and fix refactoring bugs
